refactor(frontend): rename action, reducer and util modules
