+++ server/zanata-frontend/src/frontend/app/actions/common-actions.js
