+++ server/zanata-frontend/src/frontend/app/actions/explore-actions.js
