+++ server/zanata-frontend/src/frontend/app/actions/glossary-actions.js
