+++ server/zanata-frontend/src/frontend/app/actions/languages-actions.js
