+++ server/zanata-frontend/src/frontend/app/editor/actions/controls-header-actions.js
