+++ server/zanata-frontend/src/frontend/app/editor/actions/glossary-actions.js
