+++ server/zanata-frontend/src/frontend/app/editor/actions/header-actions.js
