+++ server/zanata-frontend/src/frontend/app/editor/actions/key-shortcuts-actions.js
