+++ server/zanata-frontend/src/frontend/app/editor/actions/phrase-navigation-actions.js
