+++ server/zanata-frontend/src/frontend/app/editor/actions/phrases-actions.js
