+++ server/zanata-frontend/src/frontend/app/editor/actions/suggestions-actions.js
