+++ server/zanata-frontend/src/frontend/app/editor/reducers/context-reducer.js
