+++ server/zanata-frontend/src/frontend/app/editor/reducers/dropdown-reducer.js
