+++ server/zanata-frontend/src/frontend/app/editor/reducers/glossary-reducer.js
