+++ server/zanata-frontend/src/frontend/app/editor/reducers/header-data-reducer.js
