+++ server/zanata-frontend/src/frontend/app/editor/reducers/phrase-reducer.js
