+++ server/zanata-frontend/src/frontend/app/editor/reducers/suggestions-reducer.js
