+++ server/zanata-frontend/src/frontend/app/editor/reducers/ui-reducer.js
