+++ server/zanata-frontend/src/frontend/app/editor/utils/doc-id-util.js
