+++ server/zanata-frontend/src/frontend/app/editor/utils/phrase-util.js
