+++ server/zanata-frontend/src/frontend/app/editor/utils/status-util.js
