+++ server/zanata-frontend/src/frontend/app/reducers/common-reducer.js
@@ -2,7 +2,7 @@ import { handleActions } from 'redux-actions'
 import {
   CLEAR_MESSAGE,
   DEFAULT_LOCALE
-} from '../actions/common'
+} from '../actions/common-actions'
 
 export default handleActions({
   [CLEAR_MESSAGE]: (state, action) => {
